# docs/openapi/doco-openapi-v1.yaml
openapi: 3.1.0
info:
  title: Doco Open API
  version: 1.0.0
  description: Doco 知识库、文档正文、块和附件开放 API。标准正文格式为 Tiptap/ProseMirror JSON。
servers:
  - url: https://example.com/api/v1
security:
  - bearerAuth: []
...
paths:
...
  /folders:
    post:
      tags: [Folders]
      operationId: createFolder
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/FolderInput' } } } }
      responses: { '201': { $ref: '#/components/responses/Success' }, '400': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' } }
  /folders/{id}:
    parameters: [{ $ref: '#/components/parameters/ResourceId' }]
    get: { tags: [Folders], operationId: getFolder, responses: { '200': { $ref: '#/components/responses/Success' }, '404': { $ref: '#/components/responses/Error' } } }
    patch:
      tags: [Folders]
      operationId: updateFolder
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/FolderInput' } } } }
      responses: { '200': { $ref: '#/components/responses/Success' }, '400': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
    delete: { tags: [Folders], operationId: deleteFolder, responses: { '204': { description: 已删除 }, '404': { $ref: '#/components/responses/Error' } } }
...
  /attachments/{id}:
    parameters: [{ $ref: '#/components/parameters/ResourceId' }]
    get: { tags: [Attachments], operationId: downloadAttachment, responses: { '200': { description: 附件二进制内容, content: { application/octet-stream: { schema: { type: string, contentEncoding: binary } } } }, '404': { $ref: '#/components/responses/Error' } } }
    delete:
      tags: [Attachments]
      operationId: deleteAttachment
      parameters: [{ in: query, name: force, schema: { type: boolean, default: false } }]
      responses: { '204': { description: 已删除 }, '409': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: 'doco_<token_id>_<secret>' }
  parameters:
    ResourceId: { in: path, name: id, required: true, schema: { type: string } }
...
    IdempotencyKey: { in: header, name: Idempotency-Key, schema: { type: string, maxLength: 128 } }
...
  headers:
    RequestId: { schema: { type: string }, description: 服务端或调用方提供的请求 ID }
...
    RateLimitLimit: { schema: { type: integer } }
    RateLimitRemaining: { schema: { type: integer } }
    RateLimitReset: { schema: { type: integer } }
...
  responses:
    Success:
      description: 成功
      headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' }, X-RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }, X-RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }, X-RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' } }
      content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
...
    Error:
      description: 错误
      headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } }
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
  schemas:
    SuccessEnvelope: { type: object, required: [data, request_id], properties: { data: {}, request_id: { type: string } } }
...
    ErrorEnvelope:
      type: object
      required: [error, request_id]
      properties:
        error: { type: object, required: [type, code, message], properties: { type: { type: string }, code: { type: string }, message: { type: string }, details: { type: object } } }
        request_id: { type: string }
...
    FolderInput: { type: object, properties: { name: { type: string }, knowledge_base_id: { type: [string, integer] }, parent_id: { type: [string, integer, 'null'] } } }
...